Update to DurationUtil.cs to correct minor display issue
In the current version when selecting a day or year as the duration, it is displayed as "1days" or "1year" instead of "1 day" or "1 year". Also when picking a duration of 1 year and 1 day is displays as "1year, 1days" instead of "1 year, 1 day".

Never made a fork or issued a pull request before so don't hurt me, but I think I changed it correctly.