Fixed checks for manned/unmanned specific celestial body progress requirement.
When using a checkType (MANNED/UNMANNED), the base check in the abstract class
ProgressCelestialBodyRequirement always fails.

The reason is that CelestialBodySubtree has IsCompleteManned and IsCompleteUnmanned members,
but those must not be checked for manned/unmanned specific tests for progress requirements.

Looking at CelestialBodySubtree.IsCompleteManned is wrong when you want to be looking at
things like CelestialBodySubtree.orbit.IsCompleteManned.